Restore ConnectionSetupPayload as abstract class
Another follow-up fix for commit 0ac54d4b. Changing ConnectionSetupPayload
from abstract class to interface is problematic when framework code
compiled against newer RSocket is used in an application on existing
version.

ConnectionSetupPayload is now an abstract class again but
DefaultConnectionSetupPayload remains in the "core" sub-package to avoid
the package cycle with "frame".

See gh-778
